services/processors/requirements.txt
insta-scrape
python-dotenv
insta-scrape
python-dotenv
tweepy
boto3
botocore